Updated all Makefile and clean.sh file with new directory structure.
Bux Fix: Build gets stuck at message “Building Avalon lmdb” and
never comes out of the message On switching between Fabric build
to Standalone build and then back to any docker builds on same repo

+++ sdk/Makefile
@@ -35,7 +35,7 @@ install:
 	pip3 install dist/*.whl
 
 clean:
-	pip3 uninstall --yes $(WHEEL_FILE)
+	if [ -f $(WHEEL_FILE) ] ; then pip3 uninstall --yes $(WHEEL_FILE); fi
 	rm -rf build deps dist *.egg-info
 	find . -iname '*.pyc' -delete
 	find . -iname '__pycache__' -delete
